Update project to angular 16 (#81)
* Update project to angular 16

+++ core/libs/connection/src/lib/multi-shape-connection-handlers/characteristic--entity.service.ts
@@ -1,10 +1,19 @@
 import {NamespacesCacheService} from '@ame/cache';
-import {DefaultCharacteristic, DefaultEntity, DefaultEnumeration, DefaultProperty, DefaultUnit, DefaultEntityValue} from '@ame/meta-model';
-import {MxGraphService, MxGraphAttributeService, MxGraphShapeOverlayService, MxGraphHelper, MxGraphVisitorHelper} from '@ame/mx-graph';
+import {
+  DefaultCharacteristic,
+  DefaultEntity,
+  DefaultEnumeration,
+  DefaultProperty,
+  DefaultUnit,
+  DefaultEntityValue,
+  DefaultStructuredValue,
+} from '@ame/meta-model';
+import {MxGraphService, MxGraphAttributeService, MxGraphShapeOverlayService, MxGraphHelper} from '@ame/mx-graph';
 import {LanguageSettingsService} from '@ame/settings-dialog';
 import {Injectable} from '@angular/core';
 import {MultiShapeConnector} from '../models';
 import {mxgraph} from 'mxgraph-factory';
+import {NotificationsService} from '@ame/shared';
 
 @Injectable({
   providedIn: 'root',
@@ -19,17 +28,26 @@ export class CharacteristicEntityConnectionHandler implements MultiShapeConnecto
     private mxGraphAttributeService: MxGraphAttributeService,
     private mxGraphShapeOverlayService: MxGraphShapeOverlayService,
     private languageSettingsService: LanguageSettingsService,
-    private namespacesCacheService: NamespacesCacheService
+    private namespacesCacheService: NamespacesCacheService,
+    private notificationsService: NotificationsService
   ) {}
 
-  public connect(parentMetaModel: DefaultCharacteristic, childMetaModel: DefaultEntity, parent: mxgraph.mxCell, child: mxgraph.mxCell) {
-    this.mxGraphAttributeService.graph.getOutgoingEdges(parent).forEach(outEdge => this.removeCells(outEdge, null));
+  connect(parentMetaModel: DefaultCharacteristic, childMetaModel: DefaultEntity, parent: mxgraph.mxCell, child: mxgraph.mxCell): void {
+    if (parentMetaModel instanceof DefaultStructuredValue) {
+      return this.notificationsService.warning({
+        title: 'Unable to connect elements',
+        message: 'StructuredValue can only contain a scalar "string-like value space" value',
+        timeout: 5000,
+      });
+    }
 
     parentMetaModel.dataType = childMetaModel;
+    this.mxGraphAttributeService.graph.getOutgoingEdges(parent).forEach(outEdge => this.removeCells(outEdge, null));
     this.mxGraphShapeOverlayService.removeOverlay(parent, MxGraphHelper.getNewShapeOverlayButton(parent));
-    // add icon when you simply connect an enumeration with an entity.
+
+    // Add icon when you simply connect an enumeration with an entity.
     if (parentMetaModel instanceof DefaultEnumeration) {
-      //TODO User should be informed if he wants to change the entity. Otherwise he deletes all values.
+      // TODO: User should be informed if he wants to change the entity, otherwise, all the values will be deleted.
       if (!parentMetaModel.createdFromEditor) {
         parentMetaModel.values = [];
       }
@@ -43,39 +61,36 @@ export class CharacteristicEntityConnectionHandler implements MultiShapeConnecto
     }
 
     if (parentMetaModel.dataType?.isComplex()) {
-      const parentIncomingEdges = this.mxGraphAttributeService.graph.getIncomingEdges(parent);
-      parentIncomingEdges.forEach(edge => {
-        const edgeSourceMetaModelElement = MxGraphHelper.getModelElement(edge.source);
-
-        if (edgeSourceMetaModelElement instanceof DefaultProperty) {
-          // remove example value for complex datatypes
-          edgeSourceMetaModelElement.exampleValue = null;
-          edge.source['configuration'].fields = MxGraphVisitorHelper.getElementProperties(
-            edgeSourceMetaModelElement,
-            this.languageSettingsService
-          );
-          MxGraphHelper.updateLabel(edge.source, this.mxGraphAttributeService.graph, this.languageSettingsService);
-        }
-      });
+      this.updateChildPropertiesLabels(parent);
     }
 
     this.mxGraphService.assignToParent(child, parent);
     this.mxGraphService.formatShapes();
   }
 
-  private removeCells(edge: mxgraph.mxCell, parent: mxgraph.mxCell) {
+  private updateChildPropertiesLabels(parent: mxgraph.mxCell): void {
+    const parentIncomingEdges = this.mxGraphAttributeService.graph.getIncomingEdges(parent);
+    parentIncomingEdges.forEach(edge => {
+      const edgeSourceMetaModelElement = MxGraphHelper.getModelElement(edge.source);
+      if (edgeSourceMetaModelElement instanceof DefaultProperty) {
+        // Remove example value for complex datatypes
+        edgeSourceMetaModelElement.exampleValue = null;
+        MxGraphHelper.updateLabel(edge.source, this.mxGraphAttributeService.graph, this.languageSettingsService);
+      }
+    });
+  }
+
+  private removeCells(edge: mxgraph.mxCell, parent: mxgraph.mxCell): void {
     const metaModel = MxGraphHelper.getModelElement(edge.target);
 
-    if (metaModel instanceof DefaultUnit) {
-      return;
-    }
+    if (metaModel instanceof DefaultUnit) return;
 
-    // remove icon if we delete the edge between enumeration and entity.
+    // Remove icon if we delete the edge between enumeration and entity.
     if (metaModel instanceof DefaultEnumeration) {
       this.mxGraphShapeOverlayService.removeComplexTypeShapeOverlays(parent);
     }
 
-    //TODO should be defined in more detail
+    // TODO: Should be defined in more details
     if (metaModel instanceof DefaultEntityValue) {
       for (const child of metaModel.children) {
         MxGraphHelper.removeRelation(metaModel, child);
@@ -83,7 +98,7 @@ export class CharacteristicEntityConnectionHandler implements MultiShapeConnecto
 
       this.mxGraphAttributeService.graph.getOutgoingEdges(edge.target).forEach(outEdge => this.removeCells(outEdge, null));
       this.mxGraphService.removeCells([edge.target]);
-      this.currentCachedFile.removeCachedElement(metaModel.aspectModelUrn);
+      this.currentCachedFile.removeElement(metaModel.aspectModelUrn);
     }
 
     const parentModel = MxGraphHelper.getModelElement(edge.source);
